x/ref: Using Go's http to get the blessing instead of mojo network services.
Closes https://github.com/vanadium/issues/issues/1271

Please note that even-though this technically will break
when Mojo starts sandboxing, there are plenty of other
place in Vanadium that use Go's net package and all of them
will break as well, so there is no need for this to be special.
When sandboxing happens, we can start wrappning our network
calls in a facade that picks between mojo and native APIs or
a Go toolchain for mojo that would replace the net package's
implementation to use mojo APIs.

Change-Id: Ib21c53ea930d5f2be487ab97afb8f3d440b0f08b