services/agent/internal/lockutil: add v1 with more flexibility
v1 introduces the mac address of the lock holder in the lock file, in
order to track whether the lock seeker and lock holder are on the same
host (if not, using the pid to determine if the process is still running
is incorrect).

Also, use the pid of the process even if the 'ps' command doesn't exist.
This gives us something to work with on systems where we can't gather
more details about the process, with potentially a higher false positive
rate for StillRunning.

Change-Id: Iac23da331dbaa9f2d19055fea90c53b95a939d31