services/device/deviced/internal/starter: add delay to resolve flakiness
One theory is that shutting down the claimable server immediately after
claiming prevents the Claim reply from being sent to the client.  Adding
a delay to verify the theory.  If it turns out to be true, I'll take a
closer look at the sequence of events triggered by a Claim, to make sure
there's no bug in the order of the shutdown triggers.  See also
https://github.com/vanadium/build/issues/58

Change-Id: I4899dde034a1e02694adafb27d1242a131939b9c