services/cluster/vkube: Add InternalOnly option
Add an option in vkube.cfg to allocate only an internal IP address for
the cluster agent. This is currently only intended to be used by tests,
but could also be useful if vkube is only ever called from within the
same network where cluster agent is running.

The main benefit of setting InternalOnly=true is that the cluster agent
starts much faster on GCE. So the tests should be faster.

Change-Id: I9da98e4721d6e98c9a55f054e2859e910a035e96