longevity tests: Client interface and two implementations.
Each syncbase instance in our longevity test will have an associated
client that will interact with it.  There will be many different kinds
of clients, that should represents real-world use cases.

This CL defines a simple interface that different clients must
implement.

It also adds two simple clients:
* Writer, which does Puts to a set of collections at a given time interval, and
* Watcher, which watches a set of collections.

The next step will be add support for clients to the controller, so that
a client will be started for each instance when the tests run.

Change-Id: I804916fd37d7198a6228fdb9cedb2ee991a3f2c6