services/cluster/vkube: Add --test-namespace
Add a flag to specify which namespace to use. By default, the test still
uses a temporary one.

Change the format of the tempory namespace and bucket to include the
time when they are created. This will make garbage collection easier for
cases where the test itself isn't able to clean up.

Change-Id: I278a0f35e013473ff73d54393e73db7fe84654bb