longevity_tests: Increase time and lower expectations for number of writes.
The WriterTest starts a writer client and runs it for a period of time,
then verifies that the expected rows get written.  Then we check that
the rows match what we expect.

This is a little bit racey, because there even if the Writer is working
properly, there is no guarantee that it has actually written any rows.

I changed a few paramaters to help ensure that we get at least a few
writes:
* Writes happen every 50ms rather than 100ms.
* Wait time doubled from 1 second to 2.
* Decrease minimum expected row count from 5 to 2.

Change-Id: I2ecae6b8ed155db10000884153fffc63f268c36c