examples/fortune: Use the roaming and not the generic runtime factory
In this particular case, it means fortuned transparently remounts itself
as it moves across networks (if running on a laptop that moves around
for example) and it also turns on the debugging services.

In general though, I suspect we can simply do away with the use of the
'generic' profile. Will look into that separately.

Change-Id: Icca633576a498277351bea4ca6e1c4420076add2