longevity tests: Integrate clients into controller.
When running a universe the controller can now start a number of
clients for each device, which will perform operations on the syncbase
instance for that device.

There is a client "registry", which is used to associate names to
specific client implementations.  See the docs in control/client.go for
usage.

This required a some changes to client interface and models:
* Client interface now has Start/Stop methods, rather than just Run.
* Added Collection model, and included it in Database model.

There is also a bit of plumbing work to get databases and collections
from the model to the client.

We also fix some race conditions in the tests.

Change-Id: Id1b94ed9a11cd5d862ee88d6e6b9af4c40be83ff