longevity tests: Break controller teardown into  2 steps.
1) Stop all clients.

1.5) Sleep, to give syncbase instances time to process pending
operations (like sync)

2) Stop syncbase instances.

For the simple cases in longevity tests, we want to make sure that
syncbases have had a chance to complete all pending operations (like
syncing) before checking their consistency.  In order to do that, we
need to stop the clients first, and give the syncbases a few seconds
before stopping them as well.

In the future, we will likely want more sophisticated tests that will
kill syncbases more dangerously.  But for now we should test syncbase
under gentle conditions.

I also removed the instance.kill() method, but something like it may
come back in the future.  I also made it illegal for universe models to
"shrink", but left a todo to reconsider that choice.

Change-Id: I1e82c16287de67769fcd570fed9a46b61056b4f8