longevity tests: Integrate syncgroups.
The TestTwoDevicesWithClientsTest is a good example of what the first
longevity test will look like.  Two devices come up, each create a
database and collection.  One creates a syncgroup and the other joins.
One writes values and the other watches.

The next step will be to write a Checker that can verify both
collections end up with the same values.

Change-Id: I9527ff41bf62d4e1e524b72ff49ca4e1a69c68e1