services/device: use the local key manager in the device manager
Previously, we were using the agent-based key manager. The only benefit
to that was putting the encryption of the app keys out of the hands of
the device manager (the agent running the device manager was charged
with encrypting the key manager key store). However, we haven't really
been using the encryption feature so far, and since we're planning to
deprecate agentd, it seems sensible to simplify and use the local key
manager directly (which also has better performance than talking to the
agent key manager via the socket).

Change-Id: I5cf012316a7997c36313003c46c84209741d9c86