services/device/deviced/internal/impl: remove usage of key manager
We're going to deprecate the key manager (in favor of using one agent /
principal).  For now, change device manager to use on-disk credentials;
no agent will be used.  This change is mostly to keep the unit tests
working.  At some point we'll revisit this and start using agents.  The
principalManager abstraction will make it easier to swap different
principal management schemes in the future.

Change-Id: I975b7f11c5621e16573706d8273b3b653ef84de3