longevity tests: Checker model and simple implementation
This CL defines a model for Checkers.  It's pretty simple, with only a
Run method.

It adds a Checker called "Equality", that checks that all services have
identical databases, collections, and rows.

Checkers are integrated into controllers.  The expected usage is:

  err := Controller.Run(model)
  // ...
  err := Controller.Pause()
  err := Controller.RunChecker(equality)
  err := Controler.Resume()
  // ...

In order to implement the Equality Checker, I wrote a helper called
"DumpStream" (I'm open to better name suggestions).  DumpStream is a
stream that that returns all rows in all collections in all databases in
the service.

Change-Id: If0e271e6a609083b7c9b6aeb181ae00406af9f5a