Add vdltest package, defining vdl test types and values.
The concept is that the vdltest package provides a variety of
test types, as well as lists of test entries.  Each entry
contains a pair of values, representing a conversion test from
the Source to a value of the Target type.  Our vdl/vom tests will
start using these test cases as the basis for our conversion and
protocol conformance tests.

This is a first cut at the package.  Many TODOs remain, wrt
covering more useful cases, and better pruning out of cases that
aren't as useful.

MultiPart: 2/3
Change-Id: Ib45cc994ea7d36b4a303360d0b6611841f1f3595