services/identity: Use simple blessing files when using remote signing
Prior to this change, when using a remote signer (i.e., key held in a remote process),
the blessings to extend were specified as a full blessing store, which
included content signed by the remote key.

This was cumbersome, becuase it meant that any tweaks to the blessings
required regeneration of the blessing store, which required access to
the remote private key.

At this stage we determined that the convenience and security
benefits of not requiring the private key to update the blessings
(as we iterate on the conventions or the underlying encoding format in
VOM) outweigh the security benefit of being assured that the blessings
can only be changed by someone who has access to the private key.

Change-Id: Ib0104952f8002735616bb74c8f25072426a08560