longevity tests: Add String() methods to models.
The universe.String() is the most user-friendly, and prints all
databases, devices, users, and the topology of the universe.

For example:

    Databases:
    {database db blessing=root collections=[{collection col blessing=root}] syncgroups=[other-device/%%sync/sg]}

    Devices:
    {device writer-device-1 databases=[db] clients=[writer]}
    {device other-device databases=[db] clients=[noop]}
    {device writer-device-2 databases=[db] clients=[writer]}

    Users:
    {user Alice devices=[writer-device-1, other-device, writer-device-2]}

    Topology:
    writer-device-1 can send to [writer-device-1, other-device]
    other-device can send to [writer-device-1, other-device, writer-device-2]
    writer-device-2 can send to [other-device, writer-device-2]

This CL also gets rid of Universe.Databases and User.Databases
properties, and adds them as methods.  Otherwise they must be kept in
sync with Device.Databases.  I also got rid of the MinDatabases param,
it wasn't really being respected anyways.

Change-Id: I8ecdab5fc83111ca47108059e3173d70548490fc