runtime/internal: Enforce CloudVM invariant
InitCloudVM should always be called before CloudVMPublicAddress,
otherwise it will return the wrong result.

The rest of the code is structured in such a way that InitCloudVM should
always be called first. However, we had a instance where
CloudVMPublicAddress apparently returned the wrong result, just as if it
had been called before InitCloudVM.

For now, we will panic() if CloudVMPublicAddress is ever called first.

Change-Id: Ib569dcaf1ab63d5cdcb0c2706d11411fc398d09f