longevity tests: Support permissions on database, collection, syncgroup.
This CL adds a Permissions model, specified as a map from Tags to
UserSets.

e.g.

    p := model.Permissions{
	"Write": model.UserSet{alice},
	"Read": model.UserSet{alice, bob},
    }

Databases, Collections, and Syncgroups now have "Permissions" field.  If
this is a non-nil Permissions map, then the specified permissions will
be used to create.  If Permissions is nil, then a default permissions
object will be constructed that gives the current user all permissions
(this was the previous behavior).

With this CL, it's now possible to run tests with multiple users and
have their data sync.

Change-Id: Id3ce0bc911a9a545398fb827a4b468fdff9e83fb