New collections page in "debug browse" with row count.
Also add links to the new page from the main
Syncbase viewer page.

Eventually this Collection details page will be
expanded to include other information as specified
in go/syncbase-debug-viewer, but for now it just
has basic size information such as the count of
the number of rows in the Syncbase collection.

Change-Id: I23174acb3dc0abc6c8f4b1b5e2656649f23b7ced