lib/discovery: Compress addresses.
WARNING: This breaks backward compatibility: programs compiled
on either side of this commit may not be able to discover each other.

The most common use case of discovery is the call to AdvertiseServer,
which in turn includes all the addresses of the server in the
advertisement. There is a lot of redundant information in this case,
since every address lists out the blessing names, routing id etc.
of the server.

To cut down on space needed by the advertisement, compress the
list of addresses. And use the same technique in the BluetoothLE
plugin.

Without the changes to encoding.go, the added test cases would fail
because the addresses would be encoded in 1300 bytes (with the total
advertisement being 1443 bytes). With compression, these numbers
come down by 1000+ bytes.

Change-Id: I1a22cd6a7a7e1b8994c49deac95290eefac1d0c5