longevity tests: Syncgroups have "Creator" and "Joiner" devices.
Previously, all devices in the model would attempt to either Create or
Join a Syncgroup.  If the device was the Syncgroup Host, then it would
Create, otherwise it would join.

In the real world, not every device will be included in every Syncgroup.
For a given Syncgroup, some set of devices will create it (if it does
not yet exist), some will join it, and some will ignore it, or not know
about it.

This CL adds two properties to Syncgroup model: CreatorDevices and
JoinerDevices.  If a device is in CreatorDevices, it will attempt to
create the syncgroup.  If a devices is in JoinerDevices, it will attempt
to join the syncgroup.  Otherwise it will ignore the syncgroup.

Change-Id: I7b2abb855afe20d4bb6f5618bb6458b748be0775