allocator: add http interface
The blessing and mount point model is unchanged, hence the sb instances
cannot mount themselves.  This will change with
https://docs.google.com/document/d/1gt03Ng6mL8FQoIyss26fulNLxAUilPjxdA6PLqtQYmU/edit

The http server is only started if --http-addr is provided; so in the
absence of that, the change is a no-op (should be safe to roll out the
new binary to prod).

Change-Id: I982afce46ae33bda084e3fc67c5ac5ed2fc35a99