syncbase cgo: Add stubs for the JNI calls
The exported functions added in here are the counterpart for the
native functions added in https://v.io/c/22521. With the exception of
the JNI_OnLoad all the rest of the functions are just stubs. The
signature was obtained by first generating .h files using the javah
tool and then semi-automatically create the final form.

Change-Id: I97c92a6dae8e1a9f067cd6a96b4a637d52ded466