syncbase cgo: Properly add the 'cgo' tag
Took me some time to realize that the build tags the jni.go needs are
'(java OR android) AND cgo'. The non-flaky failures (patchset 4 for
example) from https://v.io/c/22547 were due to this mistake. The way I
bypass this issue in there was by omitting the 'cgo' tag.

Change-Id: Ib64122e4ba35af1c7d840571852c40285a673dbe