services/debug/debug: Discovery scanning support in the debug tool.
This commit introduces the "debug discovery" and "debug discovery scan"
commands to aid in debugging Vanadium services discovery. "scan"
is the only command for now, but "advertise" may come in the future.

On Linux this can only scan for services advertised over mDNS,
which is not that useful (since you have to be on the same network
and all that good stuff). But thanks to Aaron's work in
https://vanadium-review.googlesource.com/#/c/22166/
on OS X this can be used to scan advertisements from Android
and iOS over BluetoothLE as well.

These discovery commands are the first "deviation" of
the "debug" command-line tool, which started out as a
client of the "v.io/v23/services/debug" interfaces.
However, I see the "debug" tool becoming a dumping
ground for a variety of debugging/investigation
functionality instead of obscure tools other situations.
(perhaps v.io/x/ref/cmd/vrpc and v.io/x/ref/cmd/namespace
should also be folded into this?)

Change-Id: I788b5906fef7acaa14c195432debb65ce98e82d7