TBR: allocator: add more information about each instance
Creation time is extracted from deployment json.

Blessing patterns are currently reconstructed at list time (like the
mount name); TODO left to persist them when the instance is created.

Other changes:

Make it a requirement to have a namespace root set (either explicitly
via a new flag, or having the default namespace root be set).  Some of
the logic in name manipulation already assume that, and it's better to
just be able to rely on it always.

Instead of a self-signed blessing, use a blessing derived from
allocatord's own blessing when running without cluster agent.  This
ensures that the blessing is recognized by all (since it will be rooted
at dev.v.io).  To make this harmless, use a caveat that's always
'off' (Asim's suggestion).

Change-Id: Ide750c89cba979d1387ac1f355b0b931624fc5db