java: Hook and test the Service.ListDatabases and Database.Create
The listing of databases doesn't work properly so the test doesn't
attempt to list the databases after creating a database.

This change also adds the caching of appropriate class/method/field
IDs for a few classes (ArrayListClass, Id and VError). The code doing
that is isolated in the jni_lib.go file.

This change also adds a bunch of utility functions to talk to JNI and
to convert from v23_syncbase_* to Java equivalents. The code is
isolated in the jni_util.go file.

Note on V23SStringToJString: I initially wanted to write a Go unit
test but manufacturing a C.JNIEnv or bringing up an entire JVM are
both untractive options.

MultiPart: 2/2
Change-Id: I2c36eddb8015211dd67211a7dbbd700b2294aa49