allocatord: prevent stats handler from triggering oauth flow
Otherwise, the handler will set the csrf cookie, but will fail to
complete the flow via the oauth handler since the stats request is made
via ajax (so the redirect never happens).

This CL makes the distinction between handlers that want the oauth flow,
and handlers that only want to check if the user is logged in (but
proceed with an empty email address if not).  The stats handler will
return a new NotLoggedIn field in the JSON result in this case.

Change-Id: I6058d85d4d4c896c1b3751477f8bcda6e08982de