allocator: fix the debug UI
The debug UI is broken because the handler now expects both the instance
and mount name (and associated HMAC) to be present in the request.
However, links generated by the debug browser code only fill in the
mount name parameter.  Rather than change the debug browser code to pass
the instance name, this CL figures out ownership based on the mount name
and the email from the login cookie.  It expands the cache to hold all
instances belonging to a given email, and uses that cache both to look
up ownership by instance name or by mount name.  One advantage is that
once a given email's cache entry is refreshed, all their instances will
result in quick cache hits.

Independently, increase cookie validity to 1 week.

Change-Id: Id92caad1496d4519022d661aadb680dff53c554f